Convert structs to/from udt values
This commit splits the project into two subcrates:
stargate-grpc and stargate-grpc-derive.
The stargate-grpc-derive crate contains macros
for generating automatic conversions between Rust structs
and stargate_grpc::Value. Derive macros must be contained in
a separate crate because they are executed by the compiler
and not part of the library that's linked to the user code.